Add active tab and ability functionality
Now only one ability at a time is active(shows description of). When another icon is clicked on the current ability gets active class removed and the clicked on ability gets active class added. The current ability ID is tracked with a global array called activeElements, index 0 is the active tab and index 1 is the active ability. If there is not a current active the value of activeElments is "none".