Remove classPassive global elements
This was really horrible programming and I only did it to keep progressing while I was motivated. I went back and added this value to the txt file that all data comes from and included it in the bookmarks array during parsing so this global array quick fix is now gone.