Recreates passive calculations and updates ability info with these changes
Passive calculations now go through all passive abilities checking for matches to stat values in their descriptions. If a match is found the index of the match is returned and the 4 characters before the match are selected as a sub string. If this substring contains a + the value is added to the stat passive value, if not check for a - character and subtract the the value from the stat passive value. If neither +/- is within the substring add 0 to the stat value passive since this instance of matching doesn't affect character values.